refactor: fix tests
