+++ admin_panel/src/commonMain/kotlin/net/thechance/mena/admin_panel/data/repository/user/UserRepositoryImpl.kt
@@ -1,14 +1,12 @@
 package net.thechance.mena.admin_panel.data.repository.user
 
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import net.thechance.mena.admin_panel.data.mapper.toEntityList
 import net.thechance.mena.admin_panel.data.mapper.user.buildSortQuery
 import net.thechance.mena.admin_panel.data.mapper.user.toEntity
 import net.thechance.mena.admin_panel.data.remote.dto.PagedResponse
 import net.thechance.mena.admin_panel.data.remote.dto.user.UpdateUserStatusRequestDto
 import net.thechance.mena.admin_panel.data.remote.dto.user.UserResponse
-import net.thechance.mena.admin_panel.data.remote.service.UserApiService
+import net.thechance.mena.admin_panel.data.remote.api_service.UserApiService
 import net.thechance.mena.admin_panel.data.utils.executeApiSafely
 import net.thechance.mena.admin_panel.domain.entity.user.Status
 import net.thechance.mena.admin_panel.domain.entity.user.User
