+++ admin_panel/src/commonTest/kotlin/net/thechance/mena/admin_panel/data/repository/UserRepositoryImplTest.kt
@@ -11,12 +11,11 @@ import io.ktor.client.statement.HttpResponse
 import io.ktor.http.HttpStatusCode
 import io.ktor.utils.io.InternalAPI
 import junit.framework.TestCase.assertTrue
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.runTest
 import kotlinx.io.IOException
 import net.thechance.mena.admin_panel.data.remote.dto.PagedResponse
 import net.thechance.mena.admin_panel.data.remote.dto.user.UserResponse
-import net.thechance.mena.admin_panel.data.remote.service.UserApiService
+import net.thechance.mena.admin_panel.data.remote.api_service.UserApiService
 import net.thechance.mena.admin_panel.data.repository.user.UserRepositoryImpl
 import net.thechance.mena.admin_panel.domain.entity.user.Status
 import net.thechance.mena.admin_panel.domain.exceptions.NoInternetException
