+++ admin_panel/src/desktopMain/kotlin/net/thechance/mena/admin_panel/di/networkModule.desktop.kt
@@ -1,10 +1,10 @@
 package net.thechance.mena.admin_panel.di
 
 import de.jensklingenberg.ktorfit.Ktorfit
-import net.thechance.mena.admin_panel.data.remote.service.AdminAuthenticationApiService
-import net.thechance.mena.admin_panel.data.remote.service.UserApiService
-import net.thechance.mena.admin_panel.data.remote.service.createAdminAuthenticationApiService
-import net.thechance.mena.admin_panel.data.remote.service.createUserApiService
+import net.thechance.mena.admin_panel.data.remote.api_service.AdminAuthenticationApiService
+import net.thechance.mena.admin_panel.data.remote.api_service.UserApiService
+import net.thechance.mena.admin_panel.data.remote.api_service.createAdminAuthenticationApiService
+import net.thechance.mena.admin_panel.data.remote.api_service.createUserApiService
 import org.koin.core.module.Module
 import org.koin.dsl.module
 import org.koin.core.qualifier.named
